+++ test/python/multiplicative.py
@@ -7,18 +7,6 @@ from discopy.biclosed import *
 from discopy.python import *
 
 
-def test_Function():
-    x, y, z = (complex, ), (bool, ), (float, )
-    f = Function(dom=y, cod=exp(z, x),
-                 inside=lambda y: lambda x: abs(x) ** 2 if y else 0)
-    g = Function(dom=x + y, cod=z, inside=lambda x, y: f(y)(x))
-
-    assert f.uncurry().curry()(True)(1j) == f(True)(1j)
-    assert f.uncurry(left=False).curry(left=False)(True)(1j) == f(True)(1j)
-    assert g.curry().uncurry()(1j, True) == g(1j, True)
-    assert g.curry(left=False).uncurry(left=False)(1j, True) == g(1j, True)
-
-
 def test_fixed_point():
     from math import sqrt
     phi = Function(lambda x=1: 1 + 1 / x, dom=(float,), cod=(float,)).fix()
@@ -33,3 +21,17 @@ def test_trace():
 def test_list_generic_in_function():
     func = Function(sum, List[int], int)
     assert func([1, 2, 3]) == 6
+
+
+def test_strategy():
+    from hypothesis import find
+    from discopy.python.multiplicative import Function
+
+    find(Function.strategy(), lambda f: len(f.dom) >= 2 and len(f.cod) >= 2)
+
+
+def test_axioms():
+    from discopy import testing
+    from discopy.python.multiplicative import Function
+
+    testing.assert_axioms(Function)